requirements.txt
bcc>=0.18.0
rich>=10.0.0
pyelftools>=0.27
